fix: build issues and directory structure
