+++ src/app/api/auth/auth.config.ts
@@ -1,0 +1,68 @@
+import { AuthOptions } from 'next-auth';
+import { DefaultSession } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+
+declare module "next-auth" {
+  interface Session {
+    user: {
+      id: string
+    } & DefaultSession["user"]
+  }
+
+  interface JWT {
+    id: string;
+  }
+}
+
+export const authOptions: AuthOptions = {
+  providers: [ 
+    CredentialsProvider({
+      name: 'Credentials',
+      credentials: {
+        username: { label: "Username", type: "text" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        try {
+          if (!credentials?.username || !credentials?.password) {
+            throw new Error('Missing credentials');
+          }
+
+          if (credentials.username === process.env.ADMIN_USERNAME && credentials.password === process.env.ADMIN_PASSWORD) {
+            return { id: '1', name: 'Curious Frame Admin', email: 'automaton@curiouslearning.org' };
+          }
+          
+          throw new Error('Invalid credentials');
+        } catch (error) {
+          console.error('Auth error:', error);
+          return null;
+        }
+      }
+    })
+  ],
+  session: {
+    strategy: "jwt",
+    maxAge: 30 * 24 * 60 * 60, // 30 days
+    updateAge: 24 * 60 * 60, // 24 hours
+  },
+  pages: {
+    signIn: '/auth/signin',
+    error: '/auth/error',
+  },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.id = user.id;
+      }
+      return token;
+    },
+    async session({ session, token }) {
+      if (token?.id) {
+        session.user.id = token.id as string;
+      }
+      return session;
+    },
+  },
+  debug: true,
+  secret: process.env.NEXTAUTH_SECRET
+};
